- making sure attachment meta size string is set
This updates the theme to a version where our content-single-meeting.php
file checks to make sure that the `attachment_meta` value is set for a
file before it tries to determine the `filesize` parameter using it.
When it's empty it breaks the template due to a fatal error under PHP 8.
In PHP 7.4 this was simply a warning and went under the radar.