- added WP_DEBUG_LOG and SAVEQUERIES
Added the WP_DEBUG_LOG param to our standard WP configuration so that we
can turn on debugging and log it with a single constant value change in
the deploy YAML file.

Added SAVEQUERIES to use the same environment variable so that we can
set that single item to TRUE and then have all the debugging values set
as expected so we have the least amount of impact visually on the site
for users.