- preloading cache files
Given issues like [PCD108](https://github.com/proudcity/pc-dev-issues/issues/108#issuecomment-2666977511) where we see sites without CSS loaded I wonder if the database is thinking it has the cache files but we're on a new pod that just spun up and thus doesn't have the cache files in the location on the server. The server files are pod specific whereas the database is shared amongst pods behind REDIS.

This preloads the cache files when we spin up a pod.